+++ src/main/java/com/len/ticketing/application/queue/QueueAdvanceEngine.java
@@ -1,6 +1,6 @@
 package com.len.ticketing.application.queue;
 
 public interface QueueAdvanceEngine {
-    int advance(long scheduleId, long nowMs, int capacity, long passTtlSeconds);
+    int advance(long scheduleId, long nowMs, int capacity, int passTtlSeconds);
     String name(); // "lua" | "java"
 }
